refactor: combine access change modals into single component
